+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [v0.23.0] - 2020-08-27
+### Added
+- mongo: implements `storage.Transactional`
+
+### Changed
+- deps: upgrades to `mongo-driver@v1.3.7`
+
 ## [v0.22.2] - 2020-07-06
 ### Fixed
 - mongo: fixes `UserManager.Migrate` returning not found on a successful insert.
@@ -529,6 +536,7 @@ clear out the password field before sending the response.
 - General pre-release!
 
 [Unreleased]: https://github.com/matthewhartstonge/storage/tree/master
+[v0.23.0]: https://github.com/matthewhartstonge/storage/tree/v0.23.0
 [v0.22.2]: https://github.com/matthewhartstonge/storage/tree/v0.22.2
 [v0.22.1]: https://github.com/matthewhartstonge/storage/tree/v0.22.1
 [v0.22.0]: https://github.com/matthewhartstonge/storage/tree/v0.22.0